refactor: Split __init__.py and campaign.py into focused modules
Split core/views/__init__.py into focused modules:
- home.py: Dashboard views (index, account_home)
- user.py: User profile views (user, change_username)
- dice.py: Dice rolling view
- banner.py: Banner interaction views

Split core/views/campaign.py (3,263 lines) into campaign/ package:
- views.py: Class-based views (Campaigns, CampaignDetailView)
- crud.py: CRUD operations (new, edit)
- lifecycle.py: Lifecycle transitions (start, end, reopen, archive)
- lists.py: List management
- actions.py: Action logging
- assets.py: Asset management
- resources.py: Resource management
- captured.py: Captured fighter handling
- battles.py: Battle viewing
- sub_assets.py: Sub-asset CRUD
- copy.py: Campaign content copying
- common.py: Shared utilities
